+++ CHANGELOG.md
@@ -18,10 +18,12 @@ The format is based on the [KeepAChangeLog] project.
 - [#711] Deal with no post_logout_redirect_uri
 - [#712] Set Content-Type on BackChannel logout POST.
 - [#717] Missing OP logout metadata.
+- [#731] Session cookie need to be visible to OP IFrame.
 
 [#711]: https://github.com/OpenIDC/pyoidc/pull/711
 [#712]: https://github.com/OpenIDC/pyoidc/pull/712
 [#717]: https://github.com/OpenIDC/pyoidc/pull/717
+[#731]: https://github.com/OpenIDC/pyoidc/pull/731
 
 ## 1.1.1 [2019-11-04]